deploy_user_service/requirements.txt
# --- Core Web Framework ---
fastapi==0.116.1              # FastAPI web framework
uvicorn==0.35.0               # ASGI server for running FastAPI apps

# --- Database & ORM ---
sqlalchemy==2.0.41            # SQLAlchemy ORM (async supported)
asyncpg==0.30.0               # Async PostgreSQL driver

# --- Authentication & Security ---
python-jose==3.5.0            # JWT token encoding/decoding
passlib[bcrypt]==1.7.4        # Password hashing with bcrypt

# --- Environment & Settings ---
pydantic-settings==2.10.1      # Environment-based settings management (based on pydantic v2)
python-dotenv==1.1.1          # Load environment variables from .env file

# --- HTTP Requests ---
httpx==0.28.1                 # For making HTTP calls (used for MSG91 API)

# --- Testing Dependencies (Optional) ---
pytest==8.4.1                 # Test framework
pytest-asyncio==1.1.0        # Support for async test cases

pytz==2025.2          # Timezone support

gunicorn==21.2.0